DC_review/Instagram_Scraper/requirements_for_image_recog.txt
pandas
python-dotenv
groq